move telemetry to its separate package
